+++ src/main/java/at/hannibal2/skyhanni/config/features/DamageIndicator.java
@@ -55,12 +55,12 @@ public class DamageIndicator {
                     "\u00a7bDungeon Floor 6",
                     "\u00a7bDungeon Floor 7",
                     "\u00a7bDiana Mobs",
-                    "\u00a7bThunder",
+                    "\u00a7bSea Creatures",
                     "Dummy"
             }
     )
     //TODO only show currently working and tested features
-    public List<Integer> bossesToShow = new ArrayList<>(Arrays.asList(0, 1, 2, 5, 6, 7, 8, 9, 18));
+    public List<Integer> bossesToShow = new ArrayList<>(Arrays.asList(0, 1, 2, 5, 6, 7, 8, 9, 18, 19));
 
     @Expose
     @ConfigOption(name = "Hide Damage Splash", desc = "Hiding damage splashes near the damage indicator")